a bunch of small nicities
