+++ weave-js/src/components/PagePanelComponents/Home/Browse2/SpanDetails.tsx
@@ -4,11 +4,6 @@ import * as globals from '@wandb/weave/common/css/globals.styles';
 import * as _ from 'lodash';
 import React, {FC} from 'react';
 
-import {Timestamp} from '../../../Timestamp';
-import {CallStatusCodeChip} from '../Browse3/pages/common/CallStatusCodeChip';
-import {SimpleKeyValueTable} from '../Browse3/pages/common/SimplePageLayout';
-import {GroupedCalls} from '../Browse3/pages/ObjectVersionPage';
-import {WFCall} from '../Browse3/pages/wfInterface/types';
 import {Call} from './callTree';
 import {DisplayControlChars} from './CommonLib';
 import {
@@ -185,154 +180,5 @@ export const SpanDetails: FC<{
         </Box>
       </div>
     </div>
-  );
-};
-
-export const SpanDetails2: FC<{
-  wfCall: WFCall;
-}> = ({wfCall}) => {
-  const call = wfCall.rawCallSpan();
-  const childCalls = wfCall.childCalls().filter(c => {
-    return c.opVersion() != null;
-  });
-  const inputKeys =
-    call.inputs._keys ??
-    Object.entries(call.inputs)
-      .filter(([k, c]) => c != null && !k.startsWith('_'))
-      .map(([k, c]) => k);
-  const inputs = _.fromPairs(inputKeys.map(k => [k, call.inputs[k]]));
-
-  const callOutput = call.output ?? {};
-  const outputKeys =
-    callOutput._keys ??
-    Object.entries(callOutput)
-      .filter(([k, c]) => c != null && (k === '_result' || !k.startsWith('_')))
-      .map(([k, c]) => k);
-  const output = _.fromPairs(outputKeys.map(k => [k, callOutput[k]]));
-
-  const attributes = _.fromPairs(
-    Object.entries(call.attributes ?? {}).filter(
-      ([k, a]) => !k.startsWith('_') && a != null
-    )
-  );
-  const summary = _.fromPairs(
-    Object.entries(call.summary ?? {}).filter(
-      ([k, a]) => !k.startsWith('_') && k !== 'latency_s' && a != null
-    )
-  );
-
-  return (
-    <Box
-      sx={{
-        width: '100%',
-        height: '100%',
-        overflowX: 'hidden',
-        overflowY: 'auto',
-        // padding: 2,
-      }}>
-      <Box
-        sx={{
-          width: '100%',
-          display: 'flex',
-          flexDirection: 'column',
-          height: '100%',
-          gap: 1,
-          pt: 1,
-        }}>
-        <Box
-          sx={{
-            flex: '0 0 auto',
-            // border: '1px solid #ccc',
-            // borderRadius: 4,
-            p: 2,
-          }}>
-          <Typography variant="h6" gutterBottom>
-            Overview
-          </Typography>
-          <SimpleKeyValueTable
-            data={{
-              Operation:
-                parseRefMaybe(call.name) != null ? (
-                  <SmallRef
-                    objRef={parseRefMaybe(call.name)!}
-                    wfTable="OpVersion"
-                  />
-                ) : (
-                  call.name
-                ),
-              Status: (
-                <CallStatusCodeChip
-                  statusCode={call.status_code as any}
-                  showLabel
-                />
-              ),
-              Called: (
-                <Timestamp value={call.timestamp / 1000} format="relative" />
-              ),
-              ...(call.summary.latency_s != null
-                ? {
-                    Latency: (
-                      <Typography variant="body2" component="span">
-                        {call.summary.latency_s.toFixed(3)}s
-                      </Typography>
-                    ),
-                  }
-                : {}),
-              ...(call.exception ? {Exception: call.exception} : {}),
-              ...(childCalls.length > 0
-                ? {
-                    'Child Calls': (
-                      <GroupedCalls
-                        calls={childCalls}
-                        partialFilter={{
-                          parentId: wfCall.callID(),
-                        }}
-                      />
-                    ),
-                  }
-                : {}),
-              ...(Object.keys(attributes).length > 0
-                ? {Attributes: attributes}
-                : {}),
-              ...(Object.keys(summary).length > 0 ? {Summary: summary} : {}),
-            }}
-          />
-        </Box>
-        {Object.keys(inputs).length > 0 && (
-          <Box
-            sx={{
-              flex: '0 0 auto',
-              p: 2,
-            }}>
-            <Typography variant="h6" gutterBottom>
-              Inputs
-            </Typography>
-            <SimpleKeyValueTable
-              data={
-                // TODO: Consider bringing back openai chat input/output
-                inputs
-              }
-            />
-          </Box>
-        )}
-        {Object.keys(output).length > 0 && (
-          <Box
-            sx={{
-              flex: '0 0 auto',
-              p: 2,
-            }}>
-            <Typography variant="h6" gutterBottom>
-              Output
-            </Typography>
-            <SimpleKeyValueTable
-              data={
-                // TODO: Consider bringing back openai chat input/output
-                output
-              }
-            />
-          </Box>
-        )}
-      </Box>
-    </Box>
   );
 };
