+++ weave-js/src/components/PagePanelComponents/Home/Browse3/pages/common/OpVersionCategoryChip.tsx
@@ -11,6 +11,10 @@ const colorMap: {[key: string]: string} = {
   // 'tune': 'warning',
 };
 
+const ensureCapitalizedFirstLetter = (str: string) => {
+  return str.charAt(0).toUpperCase() + str.slice(1).toLowerCase();
+};
+
 export const OpVersionCategoryChip: React.FC<{
   opCategory: HackyOpCategory | null;
 }> = props => {
@@ -21,7 +25,7 @@ export const OpVersionCategoryChip: React.FC<{
   // The color seems to be borked
   return (
     <Chip
-      label={props.opCategory}
+      label={ensureCapitalizedFirstLetter(props.opCategory)}
       size="small"
       sx={{height: '20px', lineHeight: 2}}
       color={color as any}
